Fix Windows build errors and warnings
- Fix M_PI undefined: define _USE_MATH_DEFINES and M_PI for Windows
- Fix opencv2/optflow.hpp: replace with opencv2/video.hpp (OpenCV 4.x)
- Fix RotatedRect constructor: use explicit variables to avoid MSVC parsing issues
- Fix type conversion warnings: add static_cast for Eigen::Index to int
- Fix type conversion warnings: add static_cast for double to float (std::sqrt, std::pow)
- Fix type conversion warnings: add static_cast for size_t to int
- Suppress getenv warnings on MSVC with pragma directives

All Windows build errors and warnings should now be resolved.

+++ include/motcpp/motion/kalman_filters/xywh_kf.hpp
@@ -143,7 +143,7 @@ public:
         const Eigen::MatrixXf& measurements,
         bool only_position = false
     ) {
-        int n = measurements.rows();
+        int n = static_cast<int>(measurements.rows());
         Eigen::VectorXf distances(n);
         
         // Project mean and covariance
